requirements.txt
telethon
google-generativeai
requests
python-dotenv